+++ src/nullish__none_/index.d.ts
@@ -1,8 +1,12 @@
 import type { nullish } from '../_types'
-export function nullish__check_<
+export function nullish__none_<
 	T extends any[], U
 >(
 	$:T,
 	_:(...$:T)=>U,
 	onnullish?:(nullish:nullish)=>any
 ):U
+export {
+	nullish__none_ as nullish__check_,
+}
+