+++ src/nullish__not_all_/index.d.ts
@@ -1,0 +1,8 @@
+import type { nullish } from '../_types'
+export function nullish__not_all_<
+	T extends any[], U
+>(
+	$:T,
+	_:(...$:T)=>U,
+	onnullish?:(nullish:nullish)=>any
+):U